+++ doc/versionHistory.rst
@@ -5,11 +5,22 @@
 ##################
 Version History
 ##################
+
+.. _lsst.ts.wep-14.16.1:
+
+-------------
+ 14.16.1
+-------------
+
+* Fix missing logic in generateDonutFromRefitWcsTask that caused failures when scatter in WCS fit was too large.
+* Return original post_isr_image with new WCS if WCS fit successful.
+
 .. _lsst.ts.wep-14.16.0:
 
 -------------
  14.16.0
 -------------
+
 * Add DonutSizeCorrelator to measure donut diameters at the exposure level.
 * Updated handling of radius information to donut stamps metadata, and donut quality table.